+++ homework_06/app/models/__init__.py
@@ -14,11 +14,11 @@ __all__ = (
 
 from .base import Base
 from .db import db
-from .models import User
-from .models import Post
+from .user import User
+from .post import Post
 from .models import Session
 from .models import PG_CONN_URI
-from ...views.posts import posts_app
+from views.posts import posts_app
 # from .post import Post
 # from .user_profile import UserProfile
 # from .tag import Tag
