+++ homework_06/app/templates/index.html
