+++ homework_06/app/templates/posts/add.html
