+++ homework_06/app/templates/posts/details.html
