+++ homework_06/app/templates/posts/list.html
