+++ homework_06/app/templates/products/back-to-products-list.html
