+++ homework_06/app/templates/products/list.html
