+++ homework_06/app/views/__init__.py
@@ -1,0 +1,6 @@
+__all__ = (
+    "posts_app",
+
+)
+
+from .posts import posts_app
