+++ homework_06/app/views/posts.py
