+++ homework_06/app/views/products/__init__.py
