+++ homework_06/app/views/products/crud.py
