+++ homework_06/app/views/products/views.py
