+++ homework_06/docker-compose.yml
@@ -4,16 +4,16 @@ services:
   pg:
     image: postgres
     environment:
-      POSTGRES_DB: shop
+      POSTGRES_DB: blog
       POSTGRES_USER: user
       POSTGRES_PASSWORD: example
       PGDATA: /var/lib/postgresql/data
     volumes:
       - db-data:/var/lib/postgresql/data
-#    ports:
-#      - "5432:5432"
+    ports:
+      - "5432:5432"
     healthcheck:
-      test: ["CMD-SHELL", "pg_isready -d shop -U user"]
+      test: ["CMD-SHELL", "pg_isready -d blog -U user"]
       interval: 5s
       timeout: 3s
       retries: 5
@@ -30,7 +30,7 @@ services:
 #      - run
 #      - "--host=0.0.0.0"
     environment:
-      SQLALCHEMY_DATABASE_URI: "postgresql+psycopg://user:example@pg:5432/shop"
+      SQLALCHEMY_DATABASE_URI: "postgresql+psycopg://user:example@pg:5432/blog"
     ports:
       - "5000:80"
     depends_on:
